refactoring for new result that gets returned. Renamed web5 file to code-snippets
